Refactor instrument scanning and add resource manager
This commit refactors the code to include a new function `instrumentscan()` that uses the `pyvisa` library to scan for available instruments. It also adds the necessary import statement for `pyvisa` and initializes a `ResourceManager` object to manage the resources. The list of available resources is printed to the console.

The commit also includes minor changes to the `Report` class, adding a new method `setReport()`.

No changes were made to the `Instrument` and `Measure` classes.